api: introduce logs summarization method
